updated stock

+++ composeApp/src/commonMain/kotlin/ui/stock/domain/model/Stock.kt
@@ -7,7 +7,7 @@ data class Stock(
     val stockOut: Long? = null,
     val stockBox: Long? = null,
     val stockTotal: Long? = null,
-    val unitPrice: Long? = null,
+    val unitPrice: Double? = null,
     val statusCode: String? = null,
     val statusName: String? = null,
     val date: String? = null,
